+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - added timeout to provider model loading
 - compaction slider can now be properly used on touch devices
 - optimized all files loading on renderer to avoid duplicated entries
+- added specific error message when trying to install extension that requires npm while it is not available
 
 ## [0.71.0]